Refactor multi_properties::config into separate class
datatools::multi_properties couples its C++ representation and format
specification through the inner datatools::multi_properties::config class.
This makes it difficult to extend or modify the reading/writing (e.g. to
support databases) functionality, and to decouple aspects of parsing
such as file/variant lookup. The parsing is also the most complex part
of the multi_properties code, whilst the C++ representation of a multi_properties
object is very straightforward.

Split the datatools::multi_properties::config class out into a dedicated
datatools::multi_properties_config class. Update implementation of reading/writing
implementations to no longer require friendship between multi_properties and
multi_properties_config.

Update clients of datatools::multi_properties::config to use
datatools::multi_properties_config.
