+++ source/bxdatatools_module.cmake
@@ -121,6 +121,7 @@ list(APPEND ${module_name}_MODULE_HEADERS
   ${module_include_dir}/${module_name}/memory_streambuf.h
   ${module_include_dir}/${module_name}/multi_properties.h
   ${module_include_dir}/${module_name}/multi_properties.ipp
+  ${module_include_dir}/${module_name}/multi_properties_config.h
   ${module_include_dir}/${module_name}/object_configuration_description.h
   ${module_include_dir}/${module_name}/ocd_macros.h
   ${module_include_dir}/${module_name}/ocd_driver.h
@@ -290,6 +291,7 @@ ${module_source_dir}/file_include.cc
 ${module_source_dir}/logger.cc
 ${module_source_dir}/memory_streambuf.cc
 ${module_source_dir}/multi_properties.cc
+${module_source_dir}/multi_properties_config.cc
 ${module_source_dir}/object_configuration_description.cc
 ${module_source_dir}/ocd_driver.cc
 ${module_source_dir}/ocd_utils.cc
